Labeler: Add labeler glob patterns for CI
We have a "Category: CI" label in GH issues and PRs, but there is no
labeler rule to automatically mark PRs as such.

This adds a labeler rule to automatically add the "Category: CI" label
if the PR changes _any_ file in:

  - '.circleci/*'
  - '.github/*'

If all the changes are within these files, it will _not_ be labelled.

 - '.github/ISSUE_TEMPLATE/*'
 - '.github/lsan-suppressions.txt'